+++ src/api/v2/netpols.ts
@@ -1,0 +1,20 @@
+import Debug from 'debug'
+import { Operation, OperationHandlerArray } from 'express-openapi'
+import { OpenApiRequestExt } from 'src/otomi-models'
+
+const debug = Debug('otomi:api:v2:netpols')
+
+export default function (): OperationHandlerArray {
+  const get: Operation = [
+    /* business middleware not expressible by OpenAPI documentation goes here */
+    ({ otomi }: OpenApiRequestExt, res): void => {
+      debug('getAllNetpols')
+      const v = otomi.getAllAplNetpols()
+      res.json(v)
+    },
+  ]
+  const api = {
+    get,
+  }
+  return api
+}