+++ src/gitea/connect.ts
@@ -1,13 +1,19 @@
 import axios from 'axios'
 import Debug from 'debug'
+import { cleanEnv, GIT_REPO_URL } from 'src/validators'
 
 const debug = Debug('otomi:gitea-connect')
 
+const env = cleanEnv({
+  GIT_REPO_URL,
+})
+
 // get call to the api to retrieve all the commits
-export default async function giteaCheckLatest(token: string, clusterData: any): Promise<any> {
-  const domainSuffix: string | undefined = clusterData?.cluster?.domainSuffix
-  const giteaUrl = `https://gitea.${domainSuffix}/api/v1/repos/otomi/values/commits`
-  if (domainSuffix) {
+export default async function giteaCheckLatest(token: string): Promise<any> {
+  // Extracts "http://gitea-http.gitea.svc.cluster.local:3000" or "https://gitea.<domainSuffix>"
+  const baseDomain = new URL(env.GIT_REPO_URL).origin
+  if (baseDomain) {
+    const giteaUrl = `${baseDomain}/api/v1/repos/otomi/values/commits`
     const response = await axios({
       url: giteaUrl,
       method: 'GET',
@@ -18,7 +24,6 @@ export default async function giteaCheckLatest(token: string, clusterData: any):
     }).catch((error) => {
       debug('Gitea error: ', error.message)
     })
-
     return response
   }
 }