+++ src/otomi-stack.test.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-empty-function */
-import { App, CodeRepo, Policies, Team, TeamConfig, User } from 'src/otomi-models'
+import { AplCodeRepoResponse, AplServiceRequest, App, CodeRepo, Team, TeamConfig, User } from 'src/otomi-models'
 import OtomiStack from 'src/otomi-stack'
 import { mockDeep } from 'jest-mock-extended'
 import { PublicUrlExists } from './error'
@@ -57,10 +57,34 @@ describe('Data validation', () => {
     // Mocking getServices() to return a list of services
     mockTeamConfigService.getServices.mockReturnValue([
       {
-        name: 'svc',
-        ingress: { domain: 'a.com', subdomain: 'b' },
+        kind: 'AplTeamService',
+        metadata: {
+          name: 'svc',
+          labels: {
+            'apl.io/teamId': 'team-1',
+          },
+        },
+        spec: {
+          type: 'public',
+          domain: 'b.a.com',
+        },
+        status: {},
       },
-      { name: 'svc', ingress: { domain: 'a.com', subdomain: 'b', paths: ['/test/'] } },
+      {
+        kind: 'AplTeamService',
+        metadata: {
+          name: 'svc',
+          labels: {
+            'apl.io/teamId': 'team-1',
+          },
+        },
+        spec: {
+          type: 'public',
+          domain: 'b.a.com',
+          paths: ['/test/'],
+        },
+        status: {},
+      },
     ])
 
     // Ensure getTeamConfigService() returns our mocked TeamConfigService
@@ -71,29 +95,49 @@ describe('Data validation', () => {
   })
 
   test('should throw exception on duplicated domain', () => {
-    const svc = { name: 'svc', ingress: { domain: 'a.com', subdomain: 'b' } }
+    const svc: AplServiceRequest = {
+      kind: 'AplTeamService',
+      metadata: { name: 'svc' },
+      spec: { type: 'public', domain: 'b.a.com' },
+    }
     expect(() => otomiStack.checkPublicUrlInUse(teamId, svc)).toThrow(new PublicUrlExists())
   })
 
   test('should throw exception on duplicated url with path', () => {
-    const svc = { name: 'svc', ingress: { domain: 'a.com', subdomain: 'b', paths: ['/test/'] } }
+    const svc: AplServiceRequest = {
+      kind: 'AplTeamService',
+      metadata: { name: 'svc' },
+      spec: { type: 'public', domain: 'b.a.com', paths: ['/test/'] },
+    }
     expect(() => otomiStack.checkPublicUrlInUse(teamId, svc)).toThrow(new PublicUrlExists())
   })
 
   test('should not throw exception on unique url', () => {
-    const svc3 = { name: 'svc', ingress: { domain: 'a.com', subdomain: 'b', paths: ['/bla'] } }
-    expect(() => otomiStack.checkPublicUrlInUse(teamId, svc3)).not.toThrow()
+    const svc: AplServiceRequest = {
+      kind: 'AplTeamService',
+      metadata: { name: 'svc' },
+      spec: { type: 'public', domain: 'b.a.com', paths: ['/bla'] },
+    }
+    expect(() => otomiStack.checkPublicUrlInUse(teamId, svc)).not.toThrow()
   })
 
   test('should not throw exception when of type cluster', () => {
-    const svc = { name: 'svc', ingress: { type: 'cluster' } }
+    const svc: AplServiceRequest = {
+      kind: 'AplTeamService',
+      metadata: { name: 'svc' },
+      spec: { type: 'cluster' },
+    }
     expect(() => otomiStack.checkPublicUrlInUse(teamId, svc)).not.toThrow()
   })
 
   test('should not throw exception when editing', () => {
-    const svc = { id: 'x1', name: 'svc', ingress: { domain: 'a.com', subdomain: 'b', paths: ['/test/'] } }
-    const svc1 = { id: 'x1', name: 'svc', ingress: { domain: 'a.com', subdomain: 'c' } }
-    expect(() => otomiStack.checkPublicUrlInUse(teamId, svc1)).not.toThrow()
+    const svc: AplServiceRequest = {
+      kind: 'AplTeamService',
+      metadata: { name: 'svc' },
+      spec: { type: 'public', domain: 'c.a.com' },
+    }
+
+    expect(() => otomiStack.checkPublicUrlInUse(teamId, svc)).not.toThrow()
   })
 
   test('should create a password when password is not specified', async () => {
@@ -108,11 +152,11 @@ describe('Data validation', () => {
       netpols: [],
       settings: {} as Team,
       apps: [],
-      policies: {} as Policies,
+      policies: [],
       workloadValues: [],
     } as TeamConfig)
     await otomiStack.createTeam({ name: 'test' }, false)
-    expect(createItemSpy.mock.calls[0][1].password).not.toEqual('')
+    expect(createItemSpy.mock.calls[0][0].password).not.toEqual('')
     createItemSpy.mockRestore()
   })
 
@@ -128,12 +172,12 @@ describe('Data validation', () => {
       netpols: [],
       settings: {} as Team,
       apps: [],
-      policies: {} as Policies,
+      policies: [],
       workloadValues: [],
     } as TeamConfig)
     const myPassword = 'someAwesomePassword'
     await otomiStack.createTeam({ name: 'test', password: myPassword }, false)
-    expect(createItemSpy.mock.calls[0][1].password).toEqual(myPassword)
+    expect(createItemSpy.mock.calls[0][0].password).toEqual(myPassword)
     createItemSpy.mockRestore()
   })
 })
@@ -314,18 +358,22 @@ describe('Code repositories tests', () => {
     otomiStack.git = mockDeep<Git>()
 
     try {
-      otomiStack.repoService.createTeamConfig('demo', { name: 'demo' })
+      otomiStack.repoService.createTeamConfig({ name: 'demo' })
     } catch {
       // ignore
     }
     teamConfigService = otomiStack.repoService.getTeamConfigService('demo')
-    const codeRepo = {
-      id: '1',
-      teamId: 'demo',
-      name: 'code-1',
-      gitService: 'gitea',
-      repositoryUrl: 'https://gitea.test.com',
-    } as CodeRepo
+    const codeRepo: AplCodeRepoResponse = {
+      kind: 'AplTeamCodeRepo',
+      metadata: {
+        labels: {
+          'apl.io/teamId': 'demo',
+        },
+        name: 'code-1',
+      },
+      spec: { gitService: 'gitea', repositoryUrl: 'https://gitea.test.com' },
+      status: {},
+    }
 
     jest.spyOn(teamConfigService, 'getCodeRepo').mockReturnValue(codeRepo)
     jest.spyOn(otomiStack.git, 'deleteConfig').mockResolvedValue()
@@ -337,13 +385,18 @@ describe('Code repositories tests', () => {
 
   test('should create an internal code repository', async () => {
     const createItemSpy = jest.spyOn(teamConfigService, 'createCodeRepo').mockReturnValue({
-      teamId: 'demo',
-      name: 'code-1',
-      gitService: 'gitea',
-      repositoryUrl: 'https://gitea.test.com',
-    } as CodeRepo)
+      kind: 'AplTeamCodeRepo',
+      metadata: {
+        labels: {
+          'apl.io/teamId': 'demo',
+        },
+        name: 'code-1',
+      },
+      spec: { gitService: 'gitea', repositoryUrl: 'https://gitea.test.com' },
+      status: {},
+    })
 
-    const saveTeamCodeReposSpy = jest.spyOn(otomiStack, 'saveTeamCodeRepo').mockResolvedValue()
+    const saveTeamCodeReposSpy = jest.spyOn(otomiStack, 'saveTeamConfigItem').mockResolvedValue()
     const doDeploymentSpy = jest.spyOn(otomiStack, 'doTeamDeployment').mockResolvedValue()
 
     const codeRepo = await otomiStack.createCodeRepo('demo', {
@@ -359,12 +412,23 @@ describe('Code repositories tests', () => {
       repositoryUrl: 'https://gitea.test.com',
     })
     expect(createItemSpy).toHaveBeenCalledWith({
-      teamId: 'demo',
-      name: 'code-1',
-      gitService: 'gitea',
-      repositoryUrl: 'https://gitea.test.com',
+      kind: 'AplTeamCodeRepo',
+      metadata: {
+        name: 'code-1',
+      },
+      spec: { name: 'code-1', gitService: 'gitea', repositoryUrl: 'https://gitea.test.com' },
     })
-    expect(saveTeamCodeReposSpy).toHaveBeenCalledWith('demo', codeRepo)
+    expect(saveTeamCodeReposSpy).toHaveBeenCalledWith({
+      kind: 'AplTeamCodeRepo',
+      metadata: {
+        name: 'code-1',
+        labels: {
+          'apl.io/teamId': 'demo',
+        },
+      },
+      spec: { gitService: 'gitea', repositoryUrl: 'https://gitea.test.com' },
+      status: {},
+    })
     expect(doDeploymentSpy).toHaveBeenCalledWith('demo', expect.any(Function), false)
 
     createItemSpy.mockRestore()
@@ -373,30 +437,46 @@ describe('Code repositories tests', () => {
   })
 
   test('should get an existing internal code repository', () => {
-    const codeRepo = {
-      id: '1',
-      teamId: 'demo',
-      name: 'code-1',
-      gitService: 'gitea',
-      repositoryUrl: 'https://gitea.test.com',
-    } as CodeRepo
+    const codeRepo: AplCodeRepoResponse = {
+      kind: 'AplTeamCodeRepo',
+      metadata: {
+        name: 'code-1',
+        labels: {
+          'apl.io/teamId': 'demo',
+        },
+      },
+      spec: { gitService: 'gitea', repositoryUrl: 'https://gitea.test.com' },
+      status: {},
+    }
 
     jest.spyOn(teamConfigService, 'getCodeRepo').mockReturnValue(codeRepo)
 
     const result = otomiStack.getCodeRepo('demo', '1')
-    expect(result).toEqual(codeRepo)
+    expect(result).toEqual({
+      teamId: 'demo',
+      name: 'code-1',
+      gitService: 'gitea',
+      repositoryUrl: 'https://gitea.test.com',
+    })
   })
 
   test('should edit an existing internal code repository', async () => {
     const updateItemSpy = jest.spyOn(teamConfigService, 'updateCodeRepo').mockReturnValue({
-      id: '1',
-      teamId: 'demo',
-      name: 'code-1-updated',
-      gitService: 'gitea',
-      repositoryUrl: 'https://gitea.test.com',
-    } as CodeRepo)
+      kind: 'AplTeamCodeRepo',
+      metadata: {
+        name: 'code-1-updated',
+        labels: {
+          'apl.io/teamId': 'demo',
+        },
+      },
+      spec: {
+        gitService: 'gitea',
+        repositoryUrl: 'https://gitea.test.com',
+      },
+      status: {},
+    })
 
-    const saveTeamCodeReposSpy = jest.spyOn(otomiStack, 'saveTeamCodeRepo').mockResolvedValue()
+    const saveTeamCodeReposSpy = jest.spyOn(otomiStack, 'saveTeamConfigItem').mockResolvedValue()
     const doDeploymentSpy = jest.spyOn(otomiStack, 'doTeamDeployment').mockResolvedValue()
 
     const codeRepo = await otomiStack.editCodeRepo('demo', '1', {
@@ -407,19 +487,26 @@ describe('Code repositories tests', () => {
     })
 
     expect(codeRepo).toEqual({
-      id: '1',
       teamId: 'demo',
       name: 'code-1-updated',
       gitService: 'gitea',
       repositoryUrl: 'https://gitea.test.com',
     })
     expect(updateItemSpy).toHaveBeenCalledWith('1', {
-      teamId: 'demo',
-      name: 'code-1-updated',
-      gitService: 'gitea',
-      repositoryUrl: 'https://gitea.test.com',
+      metadata: { name: 'code-1-updated' },
+      spec: { name: 'code-1-updated', gitService: 'gitea', repositoryUrl: 'https://gitea.test.com' },
     })
-    expect(saveTeamCodeReposSpy).toHaveBeenCalledWith('demo', codeRepo)
+    expect(saveTeamCodeReposSpy).toHaveBeenCalledWith({
+      kind: 'AplTeamCodeRepo',
+      metadata: {
+        name: 'code-1-updated',
+        labels: {
+          'apl.io/teamId': 'demo',
+        },
+      },
+      spec: { gitService: 'gitea', repositoryUrl: 'https://gitea.test.com' },
+      status: {},
+    })
     expect(doDeploymentSpy).toHaveBeenCalledWith('demo', expect.any(Function), false)
 
     updateItemSpy.mockRestore()
@@ -429,7 +516,6 @@ describe('Code repositories tests', () => {
 
   test('should delete an existing internal code repository', async () => {
     const codeRepo = {
-      id: '1',
       teamId: 'demo',
       name: 'code-1',
       gitService: 'gitea',
@@ -451,13 +537,22 @@ describe('Code repositories tests', () => {
 
   test('should create an external public code repository', async () => {
     const createItemSpy = jest.spyOn(teamConfigService, 'createCodeRepo').mockReturnValue({
-      teamId: 'demo',
-      name: 'code-1',
-      gitService: 'github',
-      repositoryUrl: 'https://github.test.com',
-    } as CodeRepo)
+      kind: 'AplTeamCodeRepo',
+      metadata: {
+        name: 'code-1',
+        labels: {
+          'apl.io/teamId': 'demo',
+        },
+      },
+      spec: {
+        name: 'code-1',
+        gitService: 'github',
+        repositoryUrl: 'https://github.test.com',
+      },
+      status: {},
+    })
 
-    const saveTeamCodeReposSpy = jest.spyOn(otomiStack, 'saveTeamCodeRepo').mockResolvedValue()
+    const saveTeamCodeReposSpy = jest.spyOn(otomiStack, 'saveTeamConfigItem').mockResolvedValue()
     const doDeploymentSpy = jest.spyOn(otomiStack, 'doTeamDeployment').mockResolvedValue()
 
     const codeRepo = await otomiStack.createCodeRepo('demo', {
@@ -473,12 +568,29 @@ describe('Code repositories tests', () => {
       repositoryUrl: 'https://github.test.com',
     })
     expect(createItemSpy).toHaveBeenCalledWith({
-      teamId: 'demo',
-      name: 'code-1',
-      gitService: 'github',
-      repositoryUrl: 'https://github.test.com',
+      kind: 'AplTeamCodeRepo',
+      metadata: { name: 'code-1' },
+      spec: {
+        name: 'code-1',
+        gitService: 'github',
+        repositoryUrl: 'https://github.test.com',
+      },
     })
-    expect(saveTeamCodeReposSpy).toHaveBeenCalledWith('demo', codeRepo)
+    expect(saveTeamCodeReposSpy).toHaveBeenCalledWith({
+      kind: 'AplTeamCodeRepo',
+      metadata: {
+        name: 'code-1',
+        labels: {
+          'apl.io/teamId': 'demo',
+        },
+      },
+      spec: {
+        name: 'code-1',
+        gitService: 'github',
+        repositoryUrl: 'https://github.test.com',
+      },
+      status: {},
+    })
     expect(doDeploymentSpy).toHaveBeenCalledWith('demo', expect.any(Function), false)
 
     createItemSpy.mockRestore()
@@ -487,30 +599,50 @@ describe('Code repositories tests', () => {
   })
 
   test('should get an existing external public code repository', () => {
-    const codeRepo = {
-      id: '1',
-      teamId: 'demo',
-      name: 'code-1',
-      gitService: 'github',
-      repositoryUrl: 'https://github.test.com',
-    } as CodeRepo
+    const codeRepo: AplCodeRepoResponse = {
+      kind: 'AplTeamCodeRepo',
+      metadata: {
+        name: 'code-1',
+        labels: {
+          'apl.io/teamId': 'demo',
+        },
+      },
+      spec: {
+        gitService: 'github',
+        repositoryUrl: 'https://github.test.com',
+      },
+      status: {},
+    }
 
     jest.spyOn(teamConfigService, 'getCodeRepo').mockReturnValue(codeRepo)
 
     const result = otomiStack.getCodeRepo('demo', '1')
-    expect(result).toEqual(codeRepo)
+    expect(result).toEqual({
+      teamId: 'demo',
+      name: 'code-1',
+      gitService: 'github',
+      repositoryUrl: 'https://github.test.com',
+    })
   })
 
   test('should edit an existing external public code repository', async () => {
     const updateItemSpy = jest.spyOn(teamConfigService, 'updateCodeRepo').mockReturnValue({
-      id: '1',
-      teamId: 'demo',
-      name: 'code-1-updated',
-      gitService: 'github',
-      repositoryUrl: 'https://github.test.com',
-    } as CodeRepo)
+      kind: 'AplTeamCodeRepo',
+      metadata: {
+        name: 'code-1-updated',
+        labels: {
+          'apl.io/teamId': 'demo',
+        },
+      },
+      spec: {
+        name: 'code-1-updated',
+        gitService: 'github',
+        repositoryUrl: 'https://github.test.com',
+      },
+      status: {},
+    })
 
-    const saveTeamCodeReposSpy = jest.spyOn(otomiStack, 'saveTeamCodeRepo').mockResolvedValue()
+    const saveTeamCodeReposSpy = jest.spyOn(otomiStack, 'saveTeamConfigItem').mockResolvedValue()
     const doDeploymentSpy = jest.spyOn(otomiStack, 'doTeamDeployment').mockResolvedValue()
 
     const codeRepo = await otomiStack.editCodeRepo('demo', '1', {
@@ -521,19 +653,29 @@ describe('Code repositories tests', () => {
     })
 
     expect(codeRepo).toEqual({
-      id: '1',
       teamId: 'demo',
       name: 'code-1-updated',
       gitService: 'github',
       repositoryUrl: 'https://github.test.com',
     })
     expect(updateItemSpy).toHaveBeenCalledWith('1', {
-      teamId: 'demo',
-      name: 'code-1-updated',
-      gitService: 'github',
-      repositoryUrl: 'https://github.test.com',
+      metadata: { name: 'code-1-updated' },
+      spec: {
+        name: 'code-1-updated',
+        gitService: 'github',
+        repositoryUrl: 'https://github.test.com',
+      },
     })
-    expect(saveTeamCodeReposSpy).toHaveBeenCalledWith('demo', codeRepo)
+    expect(saveTeamCodeReposSpy).toHaveBeenCalledWith({
+      kind: 'AplTeamCodeRepo',
+      metadata: { name: 'code-1-updated', labels: { 'apl.io/teamId': 'demo' } },
+      spec: {
+        name: 'code-1-updated',
+        gitService: 'github',
+        repositoryUrl: 'https://github.test.com',
+      },
+      status: {},
+    })
     expect(doDeploymentSpy).toHaveBeenCalledWith('demo', expect.any(Function), false)
 
     updateItemSpy.mockRestore()
@@ -543,7 +685,6 @@ describe('Code repositories tests', () => {
 
   test('should delete an existing external public code repository', async () => {
     const codeRepo = {
-      id: '1',
       teamId: 'demo',
       name: 'code-1',
       gitService: 'github',
@@ -565,15 +706,23 @@ describe('Code repositories tests', () => {
 
   test('should create an external private code repository', async () => {
     const createItemSpy = jest.spyOn(teamConfigService, 'createCodeRepo').mockReturnValue({
-      teamId: 'demo',
-      name: 'code-1',
-      gitService: 'github',
-      repositoryUrl: 'https://github.test.com',
-      private: true,
-      secret: 'test',
-    } as CodeRepo)
+      kind: 'AplTeamCodeRepo',
+      metadata: {
+        name: 'code-1',
+        labels: {
+          'apl.io/teamId': 'demo',
+        },
+      },
+      spec: {
+        gitService: 'github',
+        repositoryUrl: 'https://github.test.com',
+        private: true,
+        secret: 'test',
+      },
+      status: {},
+    })
 
-    const saveTeamCodeReposSpy = jest.spyOn(otomiStack, 'saveTeamCodeRepo').mockResolvedValue()
+    const saveTeamCodeReposSpy = jest.spyOn(otomiStack, 'saveTeamConfigItem').mockResolvedValue()
     const doDeploymentSpy = jest.spyOn(otomiStack, 'doTeamDeployment').mockResolvedValue()
 
     const codeRepo = await otomiStack.createCodeRepo('demo', {
@@ -593,14 +742,27 @@ describe('Code repositories tests', () => {
       secret: 'test',
     })
     expect(createItemSpy).toHaveBeenCalledWith({
-      teamId: 'demo',
-      name: 'code-1',
-      gitService: 'github',
-      repositoryUrl: 'https://github.test.com',
-      private: true,
-      secret: 'test',
+      kind: 'AplTeamCodeRepo',
+      metadata: { name: 'code-1' },
+      spec: {
+        name: 'code-1',
+        gitService: 'github',
+        repositoryUrl: 'https://github.test.com',
+        private: true,
+        secret: 'test',
+      },
     })
-    expect(saveTeamCodeReposSpy).toHaveBeenCalledWith('demo', codeRepo)
+    expect(saveTeamCodeReposSpy).toHaveBeenCalledWith({
+      kind: 'AplTeamCodeRepo',
+      metadata: { name: 'code-1', labels: { 'apl.io/teamId': 'demo' } },
+      spec: {
+        gitService: 'github',
+        repositoryUrl: 'https://github.test.com',
+        private: true,
+        secret: 'test',
+      },
+      status: {},
+    })
     expect(doDeploymentSpy).toHaveBeenCalledWith('demo', expect.any(Function), false)
 
     createItemSpy.mockRestore()
@@ -610,16 +772,24 @@ describe('Code repositories tests', () => {
 
   test('should edit an existing external private code repository', async () => {
     const updateItemSpy = jest.spyOn(teamConfigService, 'updateCodeRepo').mockReturnValue({
-      id: '1',
-      teamId: 'demo',
-      name: 'code-1-updated',
-      gitService: 'github',
-      repositoryUrl: 'https://github.test.com',
-      private: true,
-      secret: 'test',
-    } as CodeRepo)
+      kind: 'AplTeamCodeRepo',
+      metadata: {
+        name: 'code-1-updated',
+        labels: {
+          'apl.io/teamId': 'demo',
+        },
+      },
+      spec: {
+        name: 'code-1-updated',
+        gitService: 'github',
+        repositoryUrl: 'https://github.test.com',
+        private: true,
+        secret: 'test',
+      },
+      status: {},
+    })
 
-    const saveTeamCodeReposSpy = jest.spyOn(otomiStack, 'saveTeamCodeRepo').mockResolvedValue()
+    const saveTeamCodeReposSpy = jest.spyOn(otomiStack, 'saveTeamConfigItem').mockResolvedValue()
     const doDeploymentSpy = jest.spyOn(otomiStack, 'doTeamDeployment').mockResolvedValue()
 
     const codeRepo = await otomiStack.editCodeRepo('demo', '1', {
@@ -632,7 +802,6 @@ describe('Code repositories tests', () => {
     })
 
     expect(codeRepo).toEqual({
-      id: '1',
       teamId: 'demo',
       name: 'code-1-updated',
       gitService: 'github',
@@ -641,14 +810,34 @@ describe('Code repositories tests', () => {
       secret: 'test',
     })
     expect(updateItemSpy).toHaveBeenCalledWith('1', {
-      teamId: 'demo',
-      name: 'code-1-updated',
-      gitService: 'github',
-      repositoryUrl: 'https://github.test.com',
-      private: true,
-      secret: 'test',
+      metadata: {
+        name: 'code-1-updated',
+      },
+      spec: {
+        name: 'code-1-updated',
+        gitService: 'github',
+        repositoryUrl: 'https://github.test.com',
+        private: true,
+        secret: 'test',
+      },
     })
-    expect(saveTeamCodeReposSpy).toHaveBeenCalledWith('demo', codeRepo)
+    expect(saveTeamCodeReposSpy).toHaveBeenCalledWith({
+      kind: 'AplTeamCodeRepo',
+      metadata: {
+        name: 'code-1-updated',
+        labels: {
+          'apl.io/teamId': 'demo',
+        },
+      },
+      spec: {
+        name: 'code-1-updated',
+        gitService: 'github',
+        repositoryUrl: 'https://github.test.com',
+        private: true,
+        secret: 'test',
+      },
+      status: {},
+    })
     expect(doDeploymentSpy).toHaveBeenCalledWith('demo', expect.any(Function), false)
 
     updateItemSpy.mockRestore()
@@ -658,7 +847,6 @@ describe('Code repositories tests', () => {
 
   test('should delete an existing external private code repository', async () => {
     const codeRepo = {
-      id: '1',
       teamId: 'demo',
       name: 'code-1',
       gitService: 'github',